refacotr: remove eventconfig
