+++ apps/vs-agent/src/plugins/index.ts
@@ -1,3 +1,2 @@
-export * from './types'
 export * from './MessagingPlugin'
 export * from './MrtdPlugin'
