+++ apps/vs-agent/src/utils/ServerConfig.ts
@@ -1,4 +1,5 @@
-import type { VsAgentNestPlugin } from '../plugins/types'
+import type { DynamicModule, Provider, Type } from '@nestjs/common'
+import type { BaseAgentModules, Plugin, VsAgent } from '@verana-labs/vs-agent-sdk'
 import type { Express } from 'express'
 
 import { DidCommFeatureQueryOptions } from '@credo-ts/didcomm'
@@ -19,4 +20,13 @@ export interface ServerConfig {
 
 export interface DidWebServerConfig extends ServerConfig {
   baseUrl: string
+}
+
+export interface VsAgentNestPlugin {
+  name: string
+  credoPlugin?: Plugin
+  controllers?: Type<any>[]
+  providers?: Provider[]
+  imports?: DynamicModule[]
+  registerEvents?: (agent: VsAgent<BaseAgentModules>, config: ServerConfig) => void
 }
