+++ apps/vs-agent/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './EventConfig'
 export * from './ServerConfig'
 export * from './logger'
 export * from './setupAgent'
